modificacion de diseño de la funcion costos

+++ database/migrations/tenant/2025_05_27_020940_create_table_cost_basic_services.php
@@ -12,7 +12,7 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('branch_office_id');
             $table->string('name');
-            $table->decimal('value', 10, 2);
+            $table->integer('value');
             $table->date('date');
             $table->text('note')->nullable();
             $table->timestamps();
